+++ client/constants/images.ts
@@ -1,0 +1,11 @@
+
+
+import Bg1 from '@/public/assets/images/BG-01.jpg'
+import Bg2 from '@/public/assets/images/BG-02.jpg'
+import Bg3 from '@/public/assets/images/BG-03.jpg'
+import Bg5 from '@/public/assets/images/BG-05.jpg'
+import Bg6 from '@/public/assets/images/BG-06.jpg'
+import Bg7 from '@/public/assets/images/BG-07.jpg'
+import Bg8 from '@/public/assets/images/BG-08.jpg'
+
+export const backgrounds = [ Bg1, Bg2, Bg3, Bg5, Bg6, Bg7, Bg8, ]
