+++ client/pages/index.tsx
@@ -10,7 +10,7 @@ import Box from '@mui/material/Box'
 
 import Bg1 from '@/public/assets/images/BG-01.jpg'
 import Bg2 from '@/public/assets/images/BG-02.jpg'
-import Bg3 from '@/public/assets/images/BG-03.jpg'
+import Bg3 from '@/public/assets/images/BG-03.jpg' 
 import Bg5 from '@/public/assets/images/BG-05.jpg'  
 import Bg6 from '@/public/assets/images/BG-06.jpg'
 import Bg7 from '@/public/assets/images/BG-07.jpg'
